Добавляет сборку проектаÐ, разделяет проект на модули
