+++ public/index.html
@@ -1920,6 +1920,6 @@
         <button class="load-more" type="button">load more</button>
       </section>
     </main>
-    <script src="../src/main.js"></script>
+    <script src="./js/bundle.js"></script>
   </body>
 </html>
