+++ src/utils.js
@@ -1,0 +1,17 @@
+const getIntervalNum = (min, max) => Math.floor(Math.random() * (max - min)) + min;
+
+const getNewArray = (arr) => {
+  const randomNum = getIntervalNum(1, arr.length + 1);
+
+  return arr.slice(0, randomNum);
+};
+
+const getCleanContainer = (container) => {
+  container.innerHTML = ``;
+};
+
+const renderElement = (container, element) => {
+  container.insertAdjacentHTML(`beforeEnd`, element);
+};
+
+export {getIntervalNum, getNewArray, getCleanContainer, renderElement};
